requirements.txt
pandas
numpy
matplotlib
seaborn
jupyter
python-dotenv
requests
openpyxl
sqlalchemy
psycopg2-binary